+++ src/runtime/memory_allocator.cc
@@ -19,7 +19,9 @@ namespace FlexFlow {
 
 // declare Legion names
 using Legion::coord_t;
+using Legion::Machine;
 using Legion::Memory;
+using Legion::Processor;
 using Realm::RegionInstance;
 
 MemoryAllocator::MemoryAllocator(Memory _memory)
@@ -49,6 +51,22 @@ void MemoryAllocator::register_reserved_work_space(void *base, size_t size) {
   reserved_ptr = base;
   reserved_total_size = size;
   reserved_allocated_size = 0;
+}
+
+Memory get_proc_mem(Machine machine, Processor proc) {
+  // First try to allocate a managed memory (cudaMallocManaged)
+  Machine::MemoryQuery proc_mem = Machine::MemoryQuery(machine)
+                                      .only_kind(Memory::GPU_MANAGED_MEM)
+                                      .has_affinity_to(proc);
+  if (proc_mem.count() > 0) {
+    return proc_mem.first();
+  }
+  // If managed memory is not available, try to allocate a framebuffer memory
+  proc_mem = Machine::MemoryQuery(machine)
+                 .only_kind(Memory::GPU_FB_MEM)
+                 .has_affinity_to(proc);
+  assert(proc_mem.count() > 0);
+  return proc_mem.first();
 }
 
 }; // namespace FlexFlow
